moving storybook assets; moving subcomponents into dedicated folder
