+++ packages/react-components/src/components/Modal/components/ModalBase.tsx
@@ -2,9 +2,8 @@ import * as React from 'react';
 
 import cx from 'clsx';
 
-import { KeyCodes } from '../../utils/keyCodes';
-
-import styles from './Modal.module.scss';
+import { KeyCodes } from '../../../utils/keyCodes';
+import styles from '../Modal.module.scss';
 
 export interface ModalBaseProps extends React.HTMLAttributes<HTMLDivElement> {
   /**
