+++ packages/react-components/src/components/Modal/components/ModalHeader.module.scss
