+++ packages/react-components/src/components/Modal/index.ts
@@ -1,11 +1,11 @@
 export { Modal } from './Modal';
-export { ModalBase } from './ModalBase';
-export { ModalCloseButton } from './ModalCloseButton';
-export { ModalPortal } from './ModalPortal';
-export { ModalHeader } from './ModalHeader';
+export { ModalBase } from './components/ModalBase';
+export { ModalCloseButton } from './components/ModalCloseButton';
+export { ModalPortal } from './components/ModalPortal';
+export { ModalHeader } from './components/ModalHeader';
 
 export type { ModalProps } from './Modal';
-export type { ModalBaseProps } from './ModalBase';
-export type { ModalCloseButtonProps } from './ModalCloseButton';
-export type { ModalPortalProps } from './ModalPortal';
-export type { ModalHeaderProps } from './ModalHeader';
+export type { ModalBaseProps } from './components/ModalBase';
+export type { ModalCloseButtonProps } from './components/ModalCloseButton';
+export type { ModalPortalProps } from './components/ModalPortal';
+export type { ModalHeaderProps } from './components/ModalHeader';
