+++ packages/react-components/src/stories/assets/modal-image.png
